# add `evaluate_test_fn` to `BaseTrainer` (Calculate the test batch performance score to determine whether it is the best model)