--- a/tianshou/data/stats.py
+++ b/tianshou/data/stats.py
@@ -56,6 +56,8 @@ class InfoStats(DataclassPPrintMixin):
 
     gradient_step: int
     """The total gradient step."""
+    best_score: float
+    """The best score over the test results."""
     best_reward: float
     """The best reward over the test results."""
     best_reward_std: float