- [ ] I have added the correct label(s) to this Pull Request or linked the relevant issue(s)
- [*] I have provided a description of the changes in this Pull Request
- [ ] I have added documentation for my changes and have listed relevant changes in CHANGELOG.md
- [ ] If applicable, I have added tests to cover my changes.
- [ ] I have reformatted the code using `poe format` 
- [ ] I have checked style and types with `poe lint` and `poe type-check`
- [ ] (Optional) I ran tests locally with `poe test` 
(or a subset of them with `poe test-reduced`) ,and they pass
- [ ] (Optional) I have tested that documentation builds correctly with `poe doc-build`

## why?
Currently, tianshou uses the maximum average return to find the best model. But sometimes it may not meet user needs, for example, the average return only drops by 5%, but the standard deviation drops by 50%. The latter is generally considered to be more stable and better than the former. So I added the `evaluate_test_fn` function, which allows users to calculate the test score by themselves, and takes the model with the largest test score as the best model. If not provided, the maximum average return is used by default.
```log
Epoch #1: 141657it [25:51, 91.32it/s, env_step=141657, gradient_step=14166, len=8, n/ep=1000, n/st=8766, rew=-0.00]
2024-08-12 14:02:31,453 INFO save best: /home/abc/saved/policy_RainbowDQN.pth
Epoch #1: test_reward: 0.001341 ± 0.048739, best_reward: 0.001341 ± 0.048739 in #1
Epoch #2: 142646it [25:50, 91.99it/s, env_step=284303, gradient_step=28430, len=9, n/ep=1000, n/st=9251, rew=-0.00]
2024-08-12 14:30:10,310 INFO save best: /home/abc/saved/policy_RainbowDQN.pth
Epoch #2: test_reward: 0.005589 ± 0.042566, best_reward: 0.005589 ± 0.042566 in #2
Epoch #3: 138159it [25:21, 90.83it/s, env_step=422462, gradient_step=42246, len=9, n/ep=1000, n/st=9859, rew=-0.00]
Epoch #3: test_reward: 0.004855 ± 0.035949, best_reward: 0.005589 ± 0.042566 in #2
Epoch #4: 145700it [26:45, 90.73it/s, env_step=568162, gradient_step=56817, len=11, n/ep=1000, n/st=11395, rew=-0.00]
Epoch #4: test_reward: 0.005201 ± 0.025295, best_reward: 0.005589 ± 0.042566 in #2
```
The average reward of batch 4, 0.005201, is slightly lower than the 0.005589 of batch 2, but the standard deviation of batch 4, 0.025295, is significantly lower than the 0.042566 of batch 2. I can make batch 4 have a higher score than batch 2 using the following function:
```python
    def evaluate_test_fn(stat: ts.data.CollectStats):
        mean, std = stat.returns_stat.mean, stat.returns_stat.std
        return mean - std * 0.08
```