@anyongjin 2024-08-14
In essence, `Average Reward` and `Test Score` are two different things. The former represents a fixed test result indicator. The latter is a score for the test result. The scoring logic for different tasks and users may be different. For example, some consider the standard deviation and some do not.
Currently, tianshou uses `best_reawrd` for both average reward and test score. It is difficult for users to implement custom scoring logic. So I suggest that `best_reward` be used only for average reward, and `best_score` be added for test score. In this way, `best_reward` and `best_score` are two different things. If it is called `best_custom_score`, people will think that there is a system default score field, so I think it is better not to add 'custom'.

Update:
* Added explanation for InfoStats.best_score.
* Use lambda function when compute_score_fn is None to avoid multiple if-else