@opcode81 2025-03-10
@anyongjin the "terminating" part was not actually implemented. If it is the score we seek to maximize, then the early stopping criterion (`stop_fn`) should also use scores instead of mean rewards, don't you agree?
I noticed this, because I am refactoring and improving the library for v2. I will change this in the v2 branch.