@anyongjin 2025-03-11
Yes, if `best_score` is used to determine whether it is the optimal model, then `best_score` should also be passed when calling `stop_fn` instead of the previously default average return. I forgot to modify this part earlier, and it should be corrected in v2.